scripts/requirements.txt
notion-client==2.2.1
requests==2.31.0
beautifulsoup4==4.12.3
google-genai>=1.0.0
